bug fix for default list actions

+++ lib/lolita/configuration/list.rb
@@ -126,16 +126,18 @@ module Lolita
       end
 
       def set_list_attributes
-        action :edit do 
-          title ::I18n.t("lolita.shared.edit")
-          url Proc.new{|view,record| view.send(:edit_lolita_resource_path, :id => record.id)}
-        end unless actions.detect{|existing_action| existing_action.name == :edit}
+        if actions.respond_to?(:each)
+          action :edit do 
+            title ::I18n.t("lolita.shared.edit")
+            url Proc.new{|view,record| view.send(:edit_lolita_resource_path, :id => record.id)}
+          end unless actions.detect{|existing_action| existing_action.name == :edit}
 
-        action :destroy do 
-          title ::I18n.t("lolita.shared.delete")
-          url Proc.new{|view,record| view.send(:lolita_resource_path,:id => record.id)}
-          html :method => :delete, :confirm => ::I18n.t("lolita.list.confirm"), :remote => true
-        end unless actions.detect{|existing_action| existing_action.name == :destroy}
+          action :destroy do 
+            title ::I18n.t("lolita.shared.delete")
+            url Proc.new{|view,record| view.send(:lolita_resource_path,:id => record.id)}
+            html :method => :delete, :confirm => ::I18n.t("lolita.list.confirm"), :remote => true
+          end unless actions.detect{|existing_action| existing_action.name == :destroy}
+        end
       end
 
     end
